35: isolate tests for node-ignore and tests for test cases
